requirements.txt
tensorflow
tensorflow_hub
matplotlib
pillow
opencv-python
